Move importSurveyResponses unit tests to all be together
