+++ packages/meditrak-server/src/tests/routeTests/index.js
@@ -5,5 +5,4 @@
 
 export { testGetChangesCount } from './changes/get';
 export { testPostChanges } from './changes/post';
-export { testImportSurveyResponses } from './surveyResponse/import';
 export { testPutSurveyResponses } from './surveyResponse/put';
